Support specifying user agent with tools
Some tools can make a lot of requests when fetching a log, so it's
useful to be able to specify a User-Agent with contact details or other
information if the log operator wants to contact a particular user of
their log.

There are still a few hardcoded user agent strings left in places where
there isn't a good way to specify them.